mig_pg_to_neo4j/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
neo4j
psycopg2
pandas
SQLAlchemy